+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## v1.1.0
+* CloudFormation support via `config-deploy.toml`.
+* Deployable Delegat repos (both for Foundation and for SOAR) must now have a `config-deploy.toml` file
+  to describe the regions, steps and nature of the deployment.
+
 ## v1.0.0
 * First release.